pyproject.toml
[project]
name = "multi-agent-ai-system"
version = "0.1.0"
description = "Multi-agent AI system with FastAPI backend for food, travel, and shopping assistance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.6",
    "langgraph>=0.0.20",
    "serpapi>=0.1.5",
    "razorpay>=1.4.0",
    "mcp[cli]>=1.9.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "master", "mcp"]
